requirements.txt
click
colorama
flask
flask-restful
jinja2
jsonschema
Markdown
mdv
ruamel.yaml
tabulate
